requirements.txt
django
djangorestframework==3.14.0
drf-spectacular==0.24.2
django-environ
pillow
uwsgi